REF: move minisat implementation into simplesat.sat.
