--- simplesat/api.py
@@ -1,4 +1,0 @@
-# TODO Reconcile names and api.
-from .minisat import Constraint, Clause, Solver  # noqa
-from .simple_solver import SimpleSATSolver  # noqa
-from .utils import value  # noqa
