+++ simplesat/sat/__init__.py
@@ -1,0 +1,1 @@
+from .minisat import MiniSATSolver  # noqa
