+++ simplesat/sat/utils.py
